Auto-Updater And Installer In One! READ DESCRIPTION CREATORS!
CREATORS! This is what you embed in a unitypackage if you want to have the component already on the root of the avatar to hint/urge your buyer to use the antirip! Do not embed Kanna Protecc itself; only this and the component on your avatar root, unproteccted!

Very important!